requirements.txt
# GPT-OSS ARC DAPO Training Requirements

# Core ML libraries
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.20.0

# Unsloth for efficient training
unsloth[colab-new]@git+https://github.com/unslothai/unsloth.git

# ARC dataset
arc-py@git+https://github.com/xu3kev/arc-py.git

# Logging and monitoring
wandb>=0.15.0

# LoRA and quantization
peft>=0.4.0
bitsandbytes>=0.41.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
openai_harmony>=0.0.4
